+++ tests/Value/OpeningHoursTest.php
@@ -2,15 +2,17 @@
 
 namespace StadGent\Services\Test\OpeningHours\Value;
 
+use DigipolisGent\Value\ValueInterface;
 use StadGent\Services\OpeningHours\Value\DayCollection;
 use StadGent\Services\OpeningHours\Value\OpeningHours;
-use StadGent\Services\OpeningHours\Value\ValueInterface;
 use PHPUnit\Framework\TestCase;
 
 /**
  * Tests the OpeningHours value object.
  *
  * @package StadGent\Services\Test\OpeningHours\Value
+ *
+ * @covers \StadGent\Services\OpeningHours\Value\OpeningHours
  */
 class OpeningHoursTest extends TestCase
 {