+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All Notable changes to `gent/services-opening-hours` package.
 
+## [Unreleased]
+
+### Changed
+
+* DMOH-49: Changed the base value objects to those of the digipolisgent/value
+  package.
+
 ## [1.0.0]
 
 ### Added